core lib added, services refactored

+++ src/document-service/Services/DocumentService.cs
@@ -1,4 +1,5 @@
-﻿using document_service.Extensions;
+﻿using core.ServerResponse;
+using document_service.Extensions;
 using document_service.Helpers;
 using document_service.Models;
 using document_service.Models.Dtos.Requests;
@@ -16,19 +17,24 @@ public class DocumentService : IDocumentService
         _repository = repository;
     }
 
-    public async Task<IEnumerable<DocumentResponse>> GetAll()
+    public async Task<Response<IEnumerable<DocumentResponse>>> GetAll()
     {
         var documents = await _repository.GetAll();
-        return documents.ToDocumentsResponse();
+        return new SuccessResponse<IEnumerable<DocumentResponse>>(documents.ToDocumentsResponse());
     }
 
-    public async Task<DocumentResponse> GetById(Guid id)
+    public async Task<Response<DocumentResponse>> GetById(Guid id)
     {
         var document = await _repository.GetById(id);
-        return document.ToDocumentResponse();
+        if (document is not null)
+        {
+            return new SuccessResponse<DocumentResponse>(document.ToDocumentResponse());
+        }
+
+        return new ErrorResponse<DocumentResponse>(ResponseStatus.NotFound, default, ResultMessage.NotFoundDocument);
     }
 
-    public async Task<string> Create(CreateDocumentRequest request)
+    public async Task<Response<string>> Create(CreateDocumentRequest request)
     {
         var fileResponse = await FileHelper.Add(request.FormFile);
         var document = new Document
@@ -38,24 +44,56 @@ public class DocumentService : IDocumentService
             Name = fileResponse.FileName,
             Description = request.LaterName
         };
-        return await _repository.Create(document);
+        var result =await _repository.Create(document);
+        if (result is not null)
+        {
+            return new SuccessResponse<string>(result);
+        }
+
+        return new ErrorResponse<string>(ResponseStatus.BadRequest,default,ResultMessage.Error);
+
+
     }
 
-    public async Task<bool> Update(Guid id, UpdateDocumentRequest request)
+    public async Task<Response<bool>> Update(Guid id, UpdateDocumentRequest request)
     {
-        var fileResponse = await FileHelper.Add(request.FormFile);
-        var document = new Document
+        var document = GetById(id);
+        if (!document.Result.Success)
         {
+            return new ErrorResponse<bool>(ResponseStatus.NotFound, default, ResultMessage.NotFoundDocument);
+        }
+        var fileResponse = await FileHelper.Add(request.FormFile);
+        var newDocument = new Document
+        {
+            Id = id,
             Extension = fileResponse.FileExtension,
             Path = fileResponse.FilePath,
             Name = fileResponse.FileName,
-            Description = request.Description
+            Description = request.Description,
+            CreatedAt = document.Result.Data.CreatedAt
         };
-        return await _repository.Update(document);
+        var result =await _repository.Update(newDocument);
+        if (result)
+        {
+            return new SuccessResponse<bool>(result);
+        }
+
+        return new ErrorResponse<bool>(result);
     }
 
-    public async Task<bool> Delete(Guid id)
+    public async Task<Response<bool>> Delete(Guid id)
     {
-        return await _repository.Delete(id);
+        var document = GetById(id);
+        if (!document.Result.Success)
+        {
+            return new ErrorResponse<bool>(ResponseStatus.NotFound, default, ResultMessage.NotFoundDocument);
+        }
+        var result = await _repository.Delete(id);
+        if (result)
+        {
+            return new SuccessResponse<bool>(result);
+        }
+
+        return new ErrorResponse<bool>(result);
     }
 }
